Split up question list into subcomponent, update snapshots
